backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
google-generativeai==0.8.2
python-dotenv==1.0.1
requests==2.31.0
PyJWT==2.9.0


